requirements.txt
aiohttp[speedups]
hikari[speedups]
hikari-lightbulb
lightbulb-ext-filament
ciso8601
python-dotenv
uvloop; sys_platform != "win32"
psutil
pillow
pytemperature
qrcode[pil]